Extract Types.hpp to break Utils.hpp <-> AbstractSyntaxTree.hpp coupling.
Move C builtin type identifiers into a shared Types.hpp so Utils no longer
pulls in the full AST header. BuiltinType keeps `using TypeId = BuiltinTypeId`
for existing call sites; Utils now depends only on Types.hpp for type IDs
and forward-declared VarType, with varTypeToTypeId still implemented in
Utils.cpp against the full AST.

+++ src/AbstractSyntaxTree.cpp
@@ -1309,7 +1309,8 @@ llvm::Value* Constant::genCode(CodeGenerator& generator) {
     case BuiltinType::TypeId::BOOL:
       return generator.getBuilder().getInt1(boolValue_);
     default:
-      throw std::logic_error("const type " + std::to_string(typeId_) +
+      throw std::logic_error("const type " +
+                             std::to_string(static_cast<int>(typeId_)) +
                              " is not supported!");
   }
 }
